+++ pyspelling/filters/html.py
@@ -11,6 +11,8 @@ import bs4
 import html
 from collections import namedtuple
 
+NON_CONTENT = (bs4.Doctype, bs4.Declaration, bs4.CData, bs4.ProcessingInstruction)
+
 RE_XML_START = re.compile(
     b'^(?:(' +
     b'<\\?xml[^>]+?>' +  # ASCII like
@@ -274,11 +276,13 @@ class HtmlFilter(filters.Filter):
         root = tree.name == '[document]'
 
         # Handle comments
-        if isinstance(tree, bs4.Comment):
+        # TODO: This might not be needed as comment objects may never get fed in this way.
+        if isinstance(tree, bs4.Comment):  # pragma: no cover
             if self.comments:
                 string = str(tree).strip()
                 if string:
-                    comments.append(html.unescape(string))
+                    sel = self.construct_selector(tree) + '<!--comment-->'
+                    comments.append((html.unescape(string), sel))
         elif root or not self.skip_tag(tree):
             # Check attributes for normal tags
             if not root:
@@ -290,6 +294,7 @@ class HtmlFilter(filters.Filter):
 
             # Walk children
             for child in tree:
+                is_comment = isinstance(child, bs4.Comment)
                 if isinstance(child, bs4.element.Tag):
                     t, b, a, c = (self.html_to_text(child))
                     text.extend(t)
@@ -297,11 +302,15 @@ class HtmlFilter(filters.Filter):
                     comments.extend(c)
                     blocks.extend(b)
                 # Get content if not the root and not a comment (unless we want comments).
-                elif not root and (not isinstance(child, bs4.Comment) or self.comments):
+                elif not isinstance(child, NON_CONTENT) and (not is_comment or self.comments):
                     string = str(child).strip()
                     if string:
-                        text.append(string)
-                        text.append(' ')
+                        if is_comment:
+                            sel = self.construct_selector(tree) + '<!--comment-->'
+                            comments.append((html.unescape(string), sel))
+                        else:
+                            text.append(string)
+                            text.append(' ')
 
         if root or self.is_block(tree):
             content = html.unescape(''.join(text))
@@ -320,8 +329,8 @@ class HtmlFilter(filters.Filter):
         content = []
         blocks, attributes, comments = self.html_to_text(bs4.BeautifulSoup(text, "html5lib"))
         if self.comments:
-            for c in comments:
-                content.append(filters.SourceText(c, context + ': <!-- comment -->', encoding, 'html-comment'))
+            for c, desc in comments:
+                content.append(filters.SourceText(c, context + ': ' + desc, encoding, 'html-comment'))
         if self.attributes:
             for a, desc in attributes:
                 content.append(filters.SourceText(a, context + ': ' + desc, encoding, 'html-attribute'))